Adding portable release.

+++ installer/exe-installer.ps1
@@ -34,8 +34,12 @@ if (-not (Test-Path $target))
     Write-Error "target folder not found at: $target"
     exit 1
 }
-Remove-Item "$target\dist" -Force -Recurse -Verbose
-Remove-Item "$target\input" -Force -Recurse -Verbose
+if (Test-Path "$target\dist") {
+    Remove-Item "$target\dist" -Force -Recurse -Verbose
+}
+if (Test-Path "$target\input") {
+    Remove-Item "$target\input" -Force -Recurse -Verbose
+}
 
 Write-Host "Copy jar to $target\input\ " -ForegroundColor Green
 mkdir "$target\input"
@@ -46,8 +50,16 @@ Write-Host "Create runtime-image with jlink: " -NoNewline -ForegroundColor Green
 Write-Host "$jLink"
 
 & $jlink `
-    --output "$target\dist\gipter-jvm" `
-    --add-modules jdk.naming.dns,jdk.management.jfr,java.rmi,jdk.jdi,java.xml,jdk.xml.dom,java.datatransfer,jdk.httpserver,java.desktop,java.security.sasl,jdk.zipfs,java.base,jdk.javadoc,jdk.management.agent,jdk.jshell,jdk.jsobject,java.sql.rowset,jdk.sctp,java.smartcardio,jdk.unsupported,java.security.jgss,java.compiler,jdk.nio.mapmode,jdk.dynalink,jdk.unsupported.desktop,jdk.accessibility,jdk.security.jgss,jdk.incubator.vector,java.sql,java.logging,java.transaction.xa,java.xml.crypto,jdk.jfr,jdk.internal.md,jdk.net,java.naming,jdk.internal.ed,java.prefs,java.net.http,jdk.compiler,jdk.internal.opt,jdk.jconsole,jdk.attach,jdk.internal.le,java.management,jdk.jdwp.agent,jdk.internal.jvmstat,java.instrument,jdk.management,jdk.security.auth,java.scripting,jdk.jartool,java.management.rmi `
+    --output "$target\dist\runtime" `
+    --add-modules jdk.naming.dns,jdk.management.jfr,java.rmi,jdk.jdi,java.xml,jdk.xml.dom,java.datatransfer `
+    --add-modules jdk.httpserver,java.desktop,java.security.sasl,jdk.zipfs,java.base,jdk.javadoc `
+    --add-modules jdk.management.agent,jdk.jshell,jdk.jsobject,java.sql.rowset,jdk.sctp,java.smartcardio `
+    --add-modules jdk.unsupported,java.security.jgss,java.compiler,jdk.nio.mapmode,jdk.dynalink `
+    --add-modules jdk.unsupported.desktop,jdk.accessibility,jdk.security.jgss,jdk.incubator.vector,java.sql `
+    --add-modules java.logging,java.transaction.xa,java.xml.crypto,jdk.jfr,jdk.internal.md,jdk.net,java.naming `
+    --add-modules jdk.internal.ed,java.prefs,java.net.http,jdk.compiler,jdk.internal.opt,jdk.jconsole,jdk.attach `
+    --add-modules jdk.internal.le,java.management,jdk.jdwp.agent,jdk.internal.jvmstat,java.instrument `
+    --add-modules jdk.management,jdk.security.auth,java.scripting,jdk.jartool,java.management.rmi `
     --strip-native-commands `
     --strip-debug `
     --no-man-pages `
@@ -59,7 +71,7 @@ Write-Host "$jPackage"
 
 & $jpackage `
     --name Gipter `
-    --runtime-image "$target/dist/gipter-jvm" `
+    --runtime-image "$target/dist/runtime" `
     --main-class pg.gipter.Java11Main `
     --main-jar "Gipter-$version.jar" `
     --dest "$target/dist" `
@@ -81,5 +93,5 @@ Write-Host "=======================================================" -Foreground
 Write-Host "jpackage DONE!" -ForegroundColor Green
 
 $ls = Get-ChildItem -Path "$currentLocation\$target\dist"
-Write-Host "Directory: $currentLocation\$target\dist"
+Write-Host "Directory: $currentLocation\$target\dist" -ForegroundColor Cyan
 Write-Host $ls
